HW -2 사용하지 않는 TextView Extension 삭제       기존 TextView Extension에서 Double형으로 꾜체
